refactor: modularize data component by extracting templates and interactions into separate files
